+++ src/Compute/Compute/ChangeLog.md
@@ -19,7 +19,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-* Allow empty value for ProximityPlacementGroupId during update
+
+## Version 3.5.0
+* Allowed empty value for ProximityPlacementGroupId during update
 
 ## Version 3.4.0
 * Limit the number of VM status to 100 to avoid throttling when Get-AzVM -Status is performed without VM name.
@@ -28,6 +30,7 @@
     - New-AzDiskUpdateConfig, New-AzSnapshotUpdateConfig
 * Add ColocationStatus parameter to Get-AzProximityPlacementGroup cmdlet.
 * Fix broken example code for 'Revoke-AzSnapshotAccess' and 'Grant-AzSnapshotAccess'
+* Fix broken example code for 'Set-AzDiskDiskEncryptionKey' and 'Set-AzDiskKeyEncryptionKey'
 
 ## Version 3.3.0
 * Fix Set-AzVMCustomScriptExtension cmdlet for a VM with managed OD disk which does not have OS profile.